+++ src/main/java/com/example/customer/contollers/CustomerController.java
@@ -1,13 +1,16 @@
 package com.example.customer.contollers;
 
 import com.example.customer.entities.Customer;
+import com.example.customer.entities.PaidType;
 import com.example.customer.services.CustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Set;
 
 @RestController
 @RequestMapping(value = "/custserv", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -17,30 +20,46 @@ public class CustomerController {
     private CustomerService customerService;
 
     @GetMapping("/customers")
-    public List<Customer> getAllCustomers(){
-        return customerService.getAllCustomers();
+    public ResponseEntity<List<Customer>> getAllCustomers(){
+        return new ResponseEntity<>(customerService.getAllCustomers(), HttpStatus.OK);
     }
 
     @GetMapping("/customers/{customerId}")
-    public Customer getCustomerByID(@PathVariable("customerId") int id){
-        return customerService.getCustomerByID(id);
+    public ResponseEntity<Customer> getCustomerByID(@PathVariable("customerId") int id){
+        if (customerService.getCustomerByID(id) == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(customerService.getCustomerByID(id), HttpStatus.OK);
+    }
+
+    @GetMapping("/customers/{customerId}/paidtypes")
+    public ResponseEntity<Set<PaidType>> getPaidTypeSet(@PathVariable("customerId") int id){
+        if (customerService.getCustomerByID(id) == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(customerService.getAllPaidTypeByCustomerID(id), HttpStatus.OK);
     }
 
     @PostMapping("/customers")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Customer addCustomer(@RequestBody Customer customer){
-        return customerService.addNewCustomer(customer);
+    public ResponseEntity<Customer> addCustomer(@RequestBody Customer customer){
+        return new ResponseEntity<>(customerService.addNewCustomer(customer), HttpStatus.CREATED);
     }
 
     @PutMapping("/customers/{customerId}")
-    public Customer updateCustomer(@PathVariable("customerId") int id,
+    public ResponseEntity<Customer> updateCustomer(@PathVariable("customerId") int id,
                                @RequestBody Customer customer){
-        return customerService.updateCustomer(id, customer);
+        if (customerService.getCustomerByID(id) == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(customerService.updateCustomer(id, customer), HttpStatus.OK);
     }
 
     @DeleteMapping("/customers/{customerId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCustomer(@PathVariable("customerId") int id){
+    public ResponseEntity<?> deleteCustomer(@PathVariable("customerId") int id){
+        if (customerService.getCustomerByID(id) == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         customerService.remove(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
